chore: rename theme switch, add hero
