+++ app/routes/resources+/theme-switch.tsx
